refactor - 컨트롤러 기능 별 분리 - 라우터 중복 유틸 분리
